Renamed buildings->building; fixed typo (#32)
